+++ release_notes.md
@@ -1,3 +1,11 @@
+MuVisionSensor V1.1.6
+==============================
+* 1.添加了球算法Label类型：`MU_BALL_TABLE_TENNIS,MU_BALL_TENNIS`
+* 2.减少了`begin`函数因通讯异常而导致的程序卡死时间
+* 3.弃用了函数`uint8_t begin(void* communication_port, MuVsMode mode)`，改为`uint8_t begin(MuVsUart* communication_port);`和`uint8_t begin(MuVsI2C* communication_port);`
+* 4.新增函数`MuVsVisionState* GetVisionState(MuVisionType vision_type)`，获取算法结果缓存地址
+* 5.添加了全新的例程、测试程序与相关文档
+
 MuVisionSensor V1.1.5
 ==============================
 * 1.修复了部分Arduino IDE无法找到头文件和examples的问题